#110 Remove search dropdown provider
