+++ src/frontend/js/modules/search/form/advanced/search-advanced-form-content.less
@@ -38,12 +38,12 @@
 
 @media (min-width: @screen-md-min) {
   .search-advanced-form-content {
-    width: 542px;
+    width: 450px;
   }
 }
 
 @media (min-width: @screen-xl-min) {
   .search-advanced-form-content {
-    width: 792px;
+    width: 700px;
   }
 }
