+++ src/frontend/js/modules/search/form/advanced/search-advanced-form.pug
@@ -1,7 +1,5 @@
 .search-advanced-form
   form.search-header-form(name='form', ng-submit='ctrl.doSearch(ctrl.searchQuery)')
-    esn-search-provider-select(ng-class="{ 'focus': focus }", providers="ctrl.providers", on-provider-selected="ctrl.onProviderSelected(provider)")
-
     .search-input-container
       button.btn.btn-icon.btn-link.button.search-header-button.search-button.hidden-xs(type="submit", title="{{ 'Search' | translate }}")
         i.mdi.mdi-magnify(ng-class="{ 'focus': focus }")
