+++ src/frontend/js/modules/search/header/search-header.less
@@ -126,8 +126,8 @@
     }
 
     .search-input {
-      /* 110px = 2 x 40px (left and right button) + margin-left button 5px + margin-right button 5px + subheader margin-left 5px + subheader margin-right 5px */
-      width: ~"calc(100vw - 166px)";
+      /* 100px = 2 x 40px (left and right button) + margin-left button 5px + margin-right button 5px + subheader margin-left 5px + subheader margin-right 5px */
+      width: ~"calc(100vw - 100px)";
     }
 
     .search-header-form {
