--- src/frontend/js/modules/search/provider-select/search-provider-select.component.js
@@ -1,17 +1,0 @@
-require('./search-provider-select.controller.js');
-
-(function(angular) {
-  'use strict';
-
-  angular
-    .module('esn.search')
-    .component('esnSearchProviderSelect', {
-      template: require('./search-provider-select.pug'),
-      controller: 'ESNSearchProviderSelectController',
-      controllerAs: 'ctrl',
-      bindings: {
-        providers: '<',
-        onProviderSelected: '&'
-      }
-    });
-})(angular);
