--- src/frontend/js/modules/search/provider-select/search-provider-select.pug
@@ -1,13 +1,0 @@
-md-select.md-no-underline(
-  ng-model="ctrl.selectedProvider",
-  ng-model-options="{trackBy: '$value.id'}"
-  ng-change="ctrl.onProviderSelected({provider: ctrl.selectedProvider})",
-  aria-label="{{ 'Search type' | translate }}",
-  title="{{ 'Search type' | translate }}",
-  md-container-class="search-provider-select-container"
-)
-  md-option.option-all(ng-value="undefined") 
-    | &nbsp;{{ 'All' | translate }}
-  md-option(ng-value="provider", ng-repeat="provider in ctrl.providers | orderBy:'displayName'")
-    img.option-icon(ng-if="provider.icon", ng-src="{{ provider.icon }}")
-    | {{ provider.displayName }}
