+++ src/frontend/js/modules/search/search.less
@@ -1,5 +1,4 @@
 @import "./header/search-header.less";
-@import "./provider-select/search-provider-select.less";
 @import "./form/advanced/search-advanced-form.less";
 @import "./form/advanced/search-advanced-form-content.less";
